Auto check PR template after submission
In order to ensure the progress tracker of each PR is representative of the actual state of the PR RSE wants to enforce that users delete the unselected checkboxes in the "Type of change" section. The checklist itself counts towards the state of the PR and is left unchanged. Users that forgot to delete checkboxes will receive an automated comment on their PR, also the template check will fail